requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
autobahn==18.4.1
Automat==0.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
psycopg2==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyOpenSSL==17.5.0
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
txaio==2.10.0
urllib3==1.22
watson-developer-cloud==1.3.0
whitenoise==3.3.1
zope.interface==4.4.3
